+++ board.py
@@ -15,3 +15,4 @@ class Board:
     
     def get_piece(self,row,col):
         return self.__positions__[row][col]
+