+++ test_main.py
@@ -1,0 +1,6 @@
+import unittest
+from test_chess import TestChess
+from test_board import TestBoard
+
+if __name__=='__main__':
+    unittest.main()