+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2024-08-15
+
+### Added
+
+-Test for board class, for get turn and the rooks creation
+-Test_main file
+
+### Known Issues
+
+- To put test in test folder, import issue
+
+
 ## [0.0.2] - 2024-08-14
 
 ### Added
 
-- Test for chess class, fro turn and change_turn methods'
+- Test for chess class, fof turn and change_turn methods'
 
 ### Known Issues